PR comments
Mostly minor debug messaging changes
Move debug messages to defines for easy change/use
Move logger to class and pass pointer to class rather than debug bool
Minor inclusion/definition errors on z/OS

+++ racfu/errors.cpp
@@ -95,7 +95,7 @@ nlohmann::json format_error_json(nlohmann::json* errors) {
   return output;
 }
 
-std::string decode_data_type(uint data_type_code) {
+std::string decode_data_type(uint8_t data_type_code) {
   switch (data_type_code) {
     case TRAIT_TYPE_BOOLEAN:
       return "a 'boolean";
